initial refactor for single meraki metricset device_health
