+++ x-pack/metricbeat/include/list.go
@@ -47,17 +47,8 @@ import (
 	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/istio/galley"
 	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/istio/mesh"
 	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/istio/mixer"
-	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/istio/pilot"
-	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/meraki"
-	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/meraki/appliance_uplink_overview"
-	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/meraki/appliance_uplink_status_and_ha"
-    _ "github.com/elastic/beats/v7/x-pack/metricbeat/module/meraki/cellular_gateway_uplink_status"
-	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/meraki/device_appliance_performance_score"
-	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/meraki/device_status"
-	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/meraki/device_uplink_loss_and_latency"
-	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/meraki/license_overview"
-	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/meraki/network_health_channel_utilization"
-	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/meraki/wireless_device_channel_utilization"
+	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/istio/pilot"	
+	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/meraki/device_health"
 	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/mssql"
 	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/mssql/performance"
 	_ "github.com/elastic/beats/v7/x-pack/metricbeat/module/mssql/transaction_log"
