--- x-pack/metricbeat/module/meraki/appliance_uplink_overview/_meta/fields.yml
@@ -1,10 +1,0 @@
-- name: appliance_uplink_overview
-  type: group
-  release: beta
-  description: >
-    appliance_uplink_overview
-  fields:
-    - name: example
-      type: keyword
-      description: >
-        Example field
