--- x-pack/metricbeat/module/meraki/appliance_uplink_status_and_ha/_meta/data.json
@@ -1,19 +1,0 @@
-{
-    "@timestamp":"2016-05-23T08:05:34.853Z",
-    "beat":{
-        "hostname":"beathost",
-        "name":"beathost"
-    },
-    "metricset":{
-        "host":"localhost",
-        "module":"meraki",
-        "name":"appliance_uplink_status_and_ha",
-        "rtt":44269
-    },
-    "meraki":{
-        "appliance_uplink_status_and_ha":{
-            "example": "appliance_uplink_status_and_ha"
-        }
-    },
-    "type":"metricsets"
-}
