--- x-pack/metricbeat/module/meraki/appliance_uplink_status_and_ha/_meta/docs.asciidoc
@@ -1,1 +1,0 @@
-This is the appliance_uplink_status_and_ha metricset of the module meraki.
