--- x-pack/metricbeat/module/meraki/appliance_uplink_status_and_ha/_meta/fields.yml
@@ -1,10 +1,0 @@
-- name: appliance_uplink_status_and_ha
-  type: group
-  release: beta
-  description: >
-    appliance_uplink_status_and_ha
-  fields:
-    - name: example
-      type: keyword
-      description: >
-        Example field
