--- x-pack/metricbeat/module/meraki/cellular_gateway_uplink_status/_meta/data.json
@@ -1,19 +1,0 @@
-{
-    "@timestamp":"2016-05-23T08:05:34.853Z",
-    "beat":{
-        "hostname":"beathost",
-        "name":"beathost"
-    },
-    "metricset":{
-        "host":"localhost",
-        "module":"meraki",
-        "name":"cellular_gateway_uplink_status",
-        "rtt":44269
-    },
-    "meraki":{
-        "cellular_gateway_uplink_status":{
-            "example": "cellular_gateway_uplink_status"
-        }
-    },
-    "type":"metricsets"
-}
