--- x-pack/metricbeat/module/meraki/cellular_gateway_uplink_status/_meta/docs.asciidoc
@@ -1,1 +1,0 @@
-This is the cellular_gateway_uplink_status metricset of the module meraki.
