--- x-pack/metricbeat/module/meraki/cellular_gateway_uplink_status/_meta/fields.yml
@@ -1,10 +1,0 @@
-- name: cellular_gateway_uplink_status
-  type: group
-  release: beta
-  description: >
-    cellular_gateway_uplink_status
-  fields:
-    - name: example
-      type: keyword
-      description: >
-        Example field
