--- x-pack/metricbeat/module/meraki/device_appliance_performance_score/_meta/data.json
@@ -1,19 +1,0 @@
-{
-    "@timestamp":"2016-05-23T08:05:34.853Z",
-    "beat":{
-        "hostname":"beathost",
-        "name":"beathost"
-    },
-    "metricset":{
-        "host":"localhost",
-        "module":"meraki",
-        "name":"device_appliance_performance_score",
-        "rtt":44269
-    },
-    "meraki":{
-        "device_appliance_performance_score":{
-            "example": "device_appliance_performance_score"
-        }
-    },
-    "type":"metricsets"
-}
