--- x-pack/metricbeat/module/meraki/device_appliance_performance_score/_meta/docs.asciidoc
@@ -1,1 +1,0 @@
-This is the device_appliance_performance_score metricset of the module meraki.
