--- x-pack/metricbeat/module/meraki/device_appliance_performance_score/_meta/fields.yml
@@ -1,10 +1,0 @@
-- name: device_appliance_performance_score
-  type: group
-  release: beta
-  description: >
-    device_appliance_performance_score
-  fields:
-    - name: example
-      type: keyword
-      description: >
-        Example field
