+++ x-pack/metricbeat/module/meraki/device_health/_meta/docs.asciidoc
@@ -1,0 +1,1 @@
+This is the device_health metricset of the module meraki.
