+++ x-pack/metricbeat/module/meraki/device_health/license_overview.go
@@ -1,94 +1,13 @@
-package license_overview
+package device_health
 
 import (
 	"fmt"
 
-	"github.com/elastic/beats/v7/libbeat/common/cfgwarn"
 	"github.com/elastic/beats/v7/metricbeat/mb"
-	"github.com/elastic/beats/v7/x-pack/metricbeat/module/meraki"
-	"github.com/elastic/elastic-agent-libs/logp"
 	"github.com/elastic/elastic-agent-libs/mapstr"
 
 	meraki_api "github.com/meraki/dashboard-api-go/v3/sdk"
 )
-
-// init registers the MetricSet with the central registry as soon as the program
-// starts. The New function will be called later to instantiate an instance of
-// the MetricSet for each host is defined in the module's configuration. After the
-// MetricSet has been created then Fetch will begin to be called periodically.
-func init() {
-	mb.Registry.MustAddMetricSet(meraki.ModuleName, "license_overview", New)
-}
-
-type config struct {
-	BaseURL       string   `config:"apiBaseURL"`
-	ApiKey        string   `config:"apiKey"`
-	DebugMode     string   `config:"apiDebugMode"`
-	Organizations []string `config:"organizations"`
-	// todo: device filtering?
-}
-
-func defaultConfig() *config {
-	return &config{
-		BaseURL:   "https://api.meraki.com",
-		DebugMode: "false",
-	}
-}
-
-// MetricSet holds any configuration or state information. It must implement
-// the mb.MetricSet interface. And this is best achieved by embedding
-// mb.BaseMetricSet because it implements all of the required mb.MetricSet
-// interface methods except for Fetch.
-type MetricSet struct {
-	mb.BaseMetricSet
-	logger        *logp.Logger
-	client        *meraki_api.Client
-	organizations []string
-}
-
-// New creates a new instance of the MetricSet. New is responsible for unpacking
-// any MetricSet specific configuration options if there are any.
-func New(base mb.BaseMetricSet) (mb.MetricSet, error) {
-	cfgwarn.Beta("The meraki license_overview metricset is beta.")
-
-	logger := logp.NewLogger(base.FullyQualifiedName())
-
-	config := defaultConfig()
-	if err := base.Module().UnpackConfig(config); err != nil {
-		return nil, err
-	}
-
-	logger.Debugf("loaded config: %v", config)
-	client, err := meraki_api.NewClientWithOptions(config.BaseURL, config.ApiKey, config.DebugMode, "Metricbeat Elastic")
-	if err != nil {
-		logger.Error("creating Meraki dashboard API client failed: %w", err)
-		return nil, err
-	}
-
-	return &MetricSet{
-		BaseMetricSet: base,
-		logger:        logger,
-		client:        client,
-		organizations: config.Organizations,
-	}, nil
-}
-
-// Fetch method implements the data gathering and data conversion to the right
-// format. It publishes the event which is then forwarded to the output. In case
-// of an error set the Error field of mb.Event or simply call report.Error().
-func (m *MetricSet) Fetch(reporter mb.ReporterV2) error {
-	for _, org := range m.organizations {
-
-		cotermLicenses, perDeviceLicenses, systemsManagerLicense, err := getLicenseStates(m.client, org)
-		if err != nil {
-			return err
-		}
-
-		reportLicenseMetrics(reporter, org, cotermLicenses, perDeviceLicenses, systemsManagerLicense)
-	}
-
-	return nil
-}
 
 func getLicenseStates(client *meraki_api.Client, organizationID string) ([]*CoterminationLicense, []*PerDeviceLicense, *SystemsManagerLicense, error) {
 	val, res, err := client.Organizations.GetOrganizationLicensesOverview(organizationID)
@@ -210,7 +129,7 @@ func reportLicenseMetrics(reporter mb.ReporterV2, organizationID string, cotermL
 				"license.count":           license.Count,
 			})
 		}
-		meraki.ReportMetricsForOrganization(reporter, organizationID, cotermLicenseMetrics)
+		ReportMetricsForOrganization(reporter, organizationID, cotermLicenseMetrics)
 	}
 
 	if len(perDeviceLicenses) != 0 {
@@ -228,12 +147,12 @@ func reportLicenseMetrics(reporter mb.ReporterV2, organizationID string, cotermL
 				"license.type":                      license.Type,
 			})
 		}
-		meraki.ReportMetricsForOrganization(reporter, organizationID, perDeviceLicenseMetrics)
+		ReportMetricsForOrganization(reporter, organizationID, perDeviceLicenseMetrics)
 	}
 
 	if systemsManagerLicense != nil {
 
-		meraki.ReportMetricsForOrganization(reporter, organizationID, []mapstr.M{
+		ReportMetricsForOrganization(reporter, organizationID, []mapstr.M{
 			{
 				"license.systems_manager.active_seats":            systemsManagerLicense.ActiveSeats,
 				"license.systems_manager.orgwideenrolled_devices": systemsManagerLicense.OrgwideEnrolledDevices,
