--- x-pack/metricbeat/module/meraki/device_status/_meta/data.json
@@ -1,19 +1,0 @@
-{
-    "@timestamp":"2016-05-23T08:05:34.853Z",
-    "beat":{
-        "hostname":"beathost",
-        "name":"beathost"
-    },
-    "metricset":{
-        "host":"localhost",
-        "module":"meraki",
-        "name":"device_status",
-        "rtt":44269
-    },
-    "meraki":{
-        "device_status":{
-            "example": "device_status"
-        }
-    },
-    "type":"metricsets"
-}
