--- x-pack/metricbeat/module/meraki/device_status/_meta/docs.asciidoc
@@ -1,1 +1,0 @@
-This is the device_status metricset of the module meraki.
