--- x-pack/metricbeat/module/meraki/device_status/_meta/fields.yml
@@ -1,10 +1,0 @@
-- name: device_status
-  type: group
-  release: beta
-  description: >
-    device_status
-  fields:
-    - name: example
-      type: keyword
-      description: >
-        Example field
