--- x-pack/metricbeat/module/meraki/device_uplink_loss_and_latency/_meta/docs.asciidoc
@@ -1,1 +1,0 @@
-This is the device_uplink_loss_and_latency metricset of the module meraki.
