--- x-pack/metricbeat/module/meraki/device_uplink_loss_and_latency/_meta/fields.yml
@@ -1,10 +1,0 @@
-- name: device_uplink_loss_and_latency
-  type: group
-  release: beta
-  description: >
-    device_uplink_loss_and_latency
-  fields:
-    - name: example
-      type: keyword
-      description: >
-        Example field
