--- x-pack/metricbeat/module/meraki/license_overview/_meta/data.json
@@ -1,19 +1,0 @@
-{
-    "@timestamp":"2016-05-23T08:05:34.853Z",
-    "beat":{
-        "hostname":"beathost",
-        "name":"beathost"
-    },
-    "metricset":{
-        "host":"localhost",
-        "module":"meraki",
-        "name":"license_overview",
-        "rtt":44269
-    },
-    "meraki":{
-        "license_overview":{
-            "example": "license_overview"
-        }
-    },
-    "type":"metricsets"
-}
