--- x-pack/metricbeat/module/meraki/license_overview/_meta/docs.asciidoc
@@ -1,1 +1,0 @@
-This is the license_overview metricset of the module meraki.
