--- x-pack/metricbeat/module/meraki/license_overview/_meta/fields.yml
@@ -1,10 +1,0 @@
-- name: license_overview
-  type: group
-  release: beta
-  description: >
-    license_overview
-  fields:
-    - name: example
-      type: keyword
-      description: >
-        Example field
