--- x-pack/metricbeat/module/meraki/network_health_channel_utilization/_meta/data.json
@@ -1,19 +1,0 @@
-{
-    "@timestamp":"2016-05-23T08:05:34.853Z",
-    "beat":{
-        "hostname":"beathost",
-        "name":"beathost"
-    },
-    "metricset":{
-        "host":"localhost",
-        "module":"meraki",
-        "name":"network_health_channel_utilization",
-        "rtt":44269
-    },
-    "meraki":{
-        "network_health_channel_utilization":{
-            "example": "network_health_channel_utilization"
-        }
-    },
-    "type":"metricsets"
-}
