--- x-pack/metricbeat/module/meraki/network_health_channel_utilization/_meta/docs.asciidoc
@@ -1,1 +1,0 @@
-This is the network_health_channel_utilization metricset of the module meraki.
