--- x-pack/metricbeat/module/meraki/wireless_device_channel_utilization/_meta/docs.asciidoc
@@ -1,1 +1,0 @@
-This is the wireless_device_channel_utilization metricset of the module meraki.
