--- x-pack/metricbeat/module/meraki/wireless_device_channel_utilization/_meta/fields.yml
@@ -1,10 +1,0 @@
-- name: wireless_device_channel_utilization
-  type: group
-  release: beta
-  description: >
-    wireless_device_channel_utilization
-  fields:
-    - name: example
-      type: keyword
-      description: >
-        Example field
