+++ x-pack/metricbeat/modules.d/meraki.yml.disabled
@@ -2,17 +2,7 @@
 # Docs: https://www.elastic.co/guide/en/beats/metricbeat/main/metricbeat-module-meraki.html
 
 - module: meraki
-  metricsets: [
-               "device_status",
-               "appliance_uplink_overview",
-               "appliance_uplink_status_and_ha",
-               "cellular_gateway_uplink_status",
-               "device_appliance_performance_score",
-               "device_uplink_loss_and_latency",
-               "license_overview",
-               "network_health_channel_utilization",
-               "wireless_device_channel_utilization"
-               ]
+  metricsets: ["device_health"]
   enabled: true
   period: 60s
   hosts: ["https://api.meraki.com"]
